requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
mistralai==0.1.3
multidict==6.0.5
numpy==1.26.4
orjson==3.9.15
pandas==2.2.1
pyarrow==15.0.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
yarl==1.9.4
